requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
alembic==1.13.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
pydantic-settings==2.5.2
passlib==1.7.4
python-jose==3.3.0
jwt==1.3.1
bcrypt==4.2.0
